added PlayPauseButton Component
