+++ src/pomodoro/TimerControls.js
@@ -1,5 +1,6 @@
 import React from "react";
-import classNames from "../utils/class-names";
+import PlayPauseButton from "./PlayPauseButton";
+import StopButton from "./StopButton";
 
 function TimerControls({
   session,
@@ -8,39 +9,6 @@ function TimerControls({
   setIsTimerRunning,
   timeRemaining,
 }) {
-  /**
-   *    Called whenever the play/pause button is clicked.
-   */
-  function playPause() {
-    setIsTimerRunning((prevState) => {
-      const nextState = !prevState;
-      if (nextState) {
-        setSession((prevStateSession) => {
-          // If the timer is starting and the previous session is null,
-          // start a focusing session.
-          if (prevStateSession === null) {
-            return {
-              label: "Focusing",
-              timeRemaining,
-            };
-          }
-          return prevStateSession;
-        });
-      }
-      return nextState;
-    });
-  }
-  /**
-   *    Called whenever the stop button is clicked
-   */
-
-  function stop() {
-    setIsTimerRunning(() => {
-      setSession(null);
-      return false;
-    });
-  }
-
   return (
     <div className="row">
       <div className="col">
@@ -49,31 +17,17 @@ function TimerControls({
           role="group"
           aria-label="Timer controls"
         >
-          <button
-            type="button"
-            className="btn btn-primary"
-            data-testid="play-pause"
-            title="Start or pause timer"
-            onClick={playPause}
-          >
-            <span
-              className={classNames({
-                oi: true,
-                "oi-media-play": !isTimerRunning,
-                "oi-media-pause": isTimerRunning,
-              })}
-            />
-          </button>
-          <button
-            type="button"
-            className="btn btn-secondary"
-            data-testid="stop"
-            title="Stop the session"
-            onClick={stop}
-            disabled={!session}
-          >
-            <span className="oi oi-media-stop" />
-          </button>
+          <PlayPauseButton
+            setSession={setSession}
+            isTimerRunning={isTimerRunning}
+            setIsTimerRunning={setIsTimerRunning}
+            timeRemaining={timeRemaining}
+          />
+          <StopButton
+            session={session}
+            setSession={setSession}
+            setIsTimerRunning={setIsTimerRunning}
+          />
         </div>
       </div>
     </div>
